+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## 0.3.2 (XXXX-XX-XX)
 
 * Add option to skip rows from the top of a csv file for the csv read rule
+* Add support for reading/writing compressed csv files for the polars backend
 
 ## 0.3.1 (2023-12-22)